Split code into main and library h/cpp
